% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amplitudes_phases}
\alias{amplitudes_phases}
\title{Amplitude and phase of seasonal forcing}
\format{A matrix with 21 rows and 3 columns:}
\usage{
amplitudes_phases
}
\description{
A dataset containing the amplitude and phase values of seasonal forcing for each patch.
}
\keyword{datasets}
