% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_brazil_age}
\alias{br_brazil_age}
\title{Brazil birth rates}
\format{
A numeric vector of length 11:
}
\usage{
br_brazil_age
}
\description{
A dataset containing the per capita birth rates for 11 age classes for Brazil.
}
\keyword{datasets}
