% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{calculate_incidence}
\alias{calculate_incidence}
\title{Calculate incidence}
\usage{
calculate_incidence(cum_infections, Ntotal, time_window)
}
\arguments{
\item{cum_infections}{array of cumulative number of infections.}

\item{Ntotal}{array of total number of individuals.}

\item{time_window}{time window over which to calculate incidence.}
}
\description{
The function calculates per capita incidence rates per 1000 individuals
}
