% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{calculate_mean_of_patch_variables}
\alias{calculate_mean_of_patch_variables}
\title{Calculate the mean across patches of a model output}
\usage{
calculate_mean_of_patch_variables(my_vector)
}
\arguments{
\item{my_vector}{vector of model outputs by patch.}
}
\description{
The function calculates the mean across patches of a model output,
by ignoring the \emph{rest of the world} patch.
}
