% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcephaly.R
\name{calculate_microcases}
\alias{calculate_microcases}
\title{Estimates total number of microcephaly cases}
\usage{
calculate_microcases(model)
}
\arguments{
\item{model}{Zika_model_simulation object.}
}
\description{
The function estimates the total number of microcephaly cases
(baseline and due to ZIKV infection during pregnancy).
}
