% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcephaly.R
\name{calculate_microcases_baseline}
\alias{calculate_microcases_baseline}
\title{Estimates baseline number of microcephaly cases}
\usage{
calculate_microcases_baseline(model, birth_rates)
}
\arguments{
\item{model}{Zika_model_simulation object.}

\item{birth_rates}{age specific birth rates for Brazil.}
}
\description{
The function estimates the baseline number of microcephaly cases
(not due to ZIKV infection during pregnancy).
}
