% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcephaly.R
\name{calculate_microcases_ZIKV}
\alias{calculate_microcases_ZIKV}
\title{Estimates number of microcephaly cases caused by Zika infection}
\usage{
calculate_microcases_ZIKV(N_inf, pregnancy_risk_curve, birth_rates)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of compartment names, e.g. \code{c("S", "R")}. In
addition a number of additional variables can be requested. These include:}
}
\description{
The function estimates the number of microcephaly cases due to ZIKV infection
during pregnancy.
}
