% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{cumsum_across_array_dims}
\alias{cumsum_across_array_dims}
\title{Cumulative sum over array margin}
\usage{
cumsum_across_array_dims(my_array, keep = NULL)
}
\arguments{
\item{my_array}{array to sum.}

\item{keep}{character vector of which variable / dimension to keep}
}
\description{
The function applies a cumulative sum over the margin of an array.
}
