% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{default_demographics}
\alias{default_demographics}
\title{Return the default demographic information}
\usage{
default_demographics()
}
\value{
list of default demographics by age group
}
\description{
Return the default demographic information
}
