% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{deterministic_model}
\alias{deterministic_model}
\title{Deterministic model creation}
\usage{
deterministic_model()
}
\description{
The function returns a list with
the odin model,
the function for creating the list of model parameters and
the function for running the model.
}
