% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{Generate equilibrium state}
\usage{
equilibrium_init_create(agec, death, model_parameter_list)
}
\arguments{
\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{model_parameter_list}{List of user-defined model parameters.}
}
\value{
Initial states of the model
}
\description{
The function creates an equilibrium initialisation state to be used within
later model runs.
}
