% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_H}
\alias{format_output_H}
\title{Format human model outputs}
\usage{
format_output_H(x, var_select = NULL, keep = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of human compartment names, e.g. \code{c("S", "R")}. In
  addition a number of additional variables can be requested. These include:
\itemize{
      \item{"infections_w"}{ Weekly Infections }
      \item{"micro_cases_w"}{ Weekly Microcephaly Cases }
      }}
}
\description{
The function formats human-realetd model outputs as a long data frame.
}
