% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_M}
\alias{format_output_M}
\title{Format mosquito model outputs}
\usage{
format_output_M(x, var_select = NULL, keep = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of mosquito compartment names, e.g. \code{c("S", "R")}. In
  addition a number of additional variables can be requested. These include:
\itemize{
      \item{"Kcav"}{ Average mosquito larvae carrying capacity by patch }
      \item{"eipav"}{ Average EIP by patch }
      \item{"Deltaav"}{ Average adult mosquito mortality rate by patch }
      \item{"R0t_1av"}{ Average Rt by patch }
      \item{"FOI1av"}{ Average FOI by patch }
      }}
}
\description{
The function formats mosquito-related model outputs as a long data frame.
}
