% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_Mprop}
\alias{format_output_Mprop}
\title{Format mosquito proportions}
\usage{
format_output_Mprop(x, keep = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{keep}{name of variable to stratify by
(only allowed \code{"patch"}. Default is no stratification)}
}
\value{
Formatted long data.frame
}
\description{
The function formats mosquito-related proportions as a long data frame.
}
