% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{lag_diff_array}
\alias{lag_diff_array}
\title{Calculate lagged differences for array}
\usage{
lag_diff_array(my_array, lag)
}
\arguments{
\item{my_array}{array of numeric values with more than 2 dimensions.}

\item{lag}{the lag value (numeric)}
}
\description{
The function returns an array of lagged difference values using \code{diff}
with initial padding
}
