% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_prob_ZIKV_pregn}
\alias{mc_prob_ZIKV_pregn}
\title{ZIKV-induced microcephaly risk during pregancy}
\format{
A dataframe of 280 rows and 2 columns:
}
\usage{
mc_prob_ZIKV_pregn
}
\description{
A dataset containing the probability that a foetus developes microcephaly
given that the mother is infected with ZIKV in a particular day of pregnancy.
}
\keyword{datasets}
