% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_parameters_list.R
\name{model_param_list_create}
\alias{model_param_list_create}
\title{model_param_list_create}
\usage{
model_param_list_create(YL = 364, DT = 1, NP = 21, age_per = 52,
  N_human = 3e+07, Omega = 1, DeltaBase = 0.2, Sigma = 0.025,
  Epsilon = 1/19, Rm = 2.69, Mwt_mean = 1.5, Mwt_cv = 0.15,
  eip_mean = 8.4, incub = 5, inf_per = 5, nu = 1/21, Kappa = 0.5,
  Beta_hm_1 = 0.7, Beta_mh_1 = 0.7, Delta_season = 1,
  Kc_season = 1, eip_season = 1, propMwtControl = 0,
  TimeMwtControlOn = 4, TimeMwtControlOff = 5, Wb_cyto = 1,
  Wb_mat = 1, Wb_fM = 0.95, Wb_fF = 0.95, Wb_relsusc1 = 0.9,
  Wb_relinf1 = 0.75, Wb_starttime = 140, Wb_introlevel = 0.5,
  Wb_introduration = 60, vacc_cu_minage = 2, vacc_cu_maxage = 15,
  vacc_cu_coverage = 0.7, vacc_cu_time = 100, vacc_child_age = 3,
  vacc_child_coverage = 0.75, vacc_child_starttime = 30,
  vacc_child_stoptime = 30, dis_pri = 0.2, rho_prim = 1,
  phi_prim = 1, vacceff_prim = 0.3, other_foi = 0.05,
  other_prop_immune = 0.5, propTransGlobal = 5e-04,
  propTransNN = 0.005, BG_FOI = 1e-06, AGE_REC = 2,
  PropDiseaseReported = 0.1)
}
\description{
\code{model_param_list_create} creates a list of user-defined
parameters
}
