% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn_links}
\alias{nn_links}
\title{Patches' nearest neighbors}
\format{A matrix with 21 rows and 8 columns:}
\usage{
nn_links
}
\description{
A matrix containing the id of the 8 nearest neighboring patches to each patch.
 The variables are as follows:
}
\references{
\url{https://science.sciencemag.org/node/681801.full}
}
\keyword{datasets}
