% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{plot.Zika_model_simulation}
\alias{plot.Zika_model_simulation}
\title{Plot human or mosquitoes states}
\usage{
\method{plot}{Zika_model_simulation}(x, type, var_select = NULL, ...)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{type}{Character of model compartment type to plot.
allowed are \code{c("H", "M")} for Human and Mosquito.}

\item{var_select}{Vector of variable names to plot (default is all).}

\item{...}{additional arguments affecting the plot produced.}
}
\description{
The function makes a plot of the model compartments.
}
