% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compartments.R
\name{plot_compartments}
\alias{plot_compartments}
\title{plot_compartments}
\usage{
plot_compartments(df, compart_names, ttl, out_pth = NULL,
  out_fl_nm = NULL)
}
\description{
\code{plot_compartments} makes a plot of the model compartments
}
