% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{plot_diagnostics}
\usage{
plot_diagnostics(df, out_pth, out_fl_nm, diagno_nms, no_pages)
}
\arguments{
\item{df}{The dataframe with the data to plot.}

\item{out_pth}{The path where to save the plot}

\item{out_fl_nm}{The output file name}

\item{diagno_nms}{A character string of descriptive names of the diagnostics. This will be used for the facet y labels.}

\item{no_pages}{The number of pages over which to arrange the diagnostics.}
}
\description{
\code{plot_diagnostics} makes a faceted plot of the model diagnostics.
}
