% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demographics.R
\name{plot_Kc_eip_delta}
\alias{plot_Kc_eip_delta}
\title{Plot seasonally-varying variables}
\usage{
plot_Kc_eip_delta(out)
}
\arguments{
\item{df}{The dataframe with the data to plot.}
}
\description{
The function makes plots of the mosquito larvae carrying capacity, virus extrinsic
 incubation period and adult mosquito mortality rate, and arrange them in a grid.
 These variables are those directly affected by seasonality in the model.
}
