% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_processing_mos}
\alias{post_processing_mos}
\title{Post-process model outputs for mosquitoes}
\usage{
post_processing_mos(dat)
}
\arguments{
\item{dat}{list of model outputs from the model run.}
}
\description{
The function processes model outputs to calculate metrics of interest for the
 vector population.
}
