% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(age, death, nn_links, time = 18200, season = FALSE)
}
\arguments{
\item{age}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{time}{time in days.
Default=18200}

\item{season}{Logical for controlling the effect of seasonality.}
}
\description{
\code{run_model} runs model using declared agec, death and nn_links
}
