% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{sum_across_array_dims}
\alias{sum_across_array_dims}
\title{Sum over aray margin}
\usage{
sum_across_array_dims(my_array, keep = NULL, compartment = NULL)
}
\arguments{
\item{my_array}{array to sum.}

\item{keep}{character vector of which variable / dimension to keep.}

\item{compartment}{character vector of compartment name.}
}
\description{
The function applies a sum over the margin of an array.
}
