% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcephaly.R
\name{calculate_microcases_ZIKV}
\alias{calculate_microcases_ZIKV}
\title{Estimates number of microcephaly cases caused by Zika infection}
\usage{
calculate_microcases_ZIKV(model, birth_rates)
}
\arguments{
\item{model}{Zika_model_simulation object.}

\item{birth_rates}{age specific birth rates for Brazil.}
}
\description{
The function estimates the number of microcephaly cases due to ZIKV infection
during pregnancy.
}
