% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_model.R
\name{create_r_model}
\alias{create_r_model}
\title{Create an odin generator function}
\usage{
create_r_model(odin_model_path, agec, death, nn_links, amplitudes_phases,
  vaccine_age = NULL, params = NULL)
}
\arguments{
\item{odin_model_path}{Character path to odin model.}

\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{amplitudes_phases}{Amplitude and phase of seasonal forcing for each patch.}

\item{vaccine_age}{Vector of binary indicators for age groups vaccination.
1 = vaccinate, 0 = do not vaccinate. Same length as \code{agec}. Default = NULL.}

\item{params}{List of user-defined parameters.
Use same names as in \code{create_model_param_list()}. Default = NULL.}
}
\value{
list of generator function, initial state, model parameters and generator
}
\description{
The function returns a list with the equilibrium initialisation state,
odin generator function automatically created given the odin model specified,
and list of user-defined parameters.
}
