% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_model.R
\name{create_r_model}
\alias{create_r_model}
\title{\code{create_r_model} returns list with generator function automatically created given the odin model
specified.}
\usage{
create_r_model(odin_model_path, agec, death, nn_links, ...)
}
\arguments{
\item{odin_model_path}{Character path to odin model}

\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}
}
\value{
list of generator function, initial state, model parameters and generator
}
\description{
\code{create_r_model} returns list with generator function automatically created given the odin model
specified.
}
