% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_init_create.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{Generate equilibrium state}
\usage{
equilibrium_init_create(agec, death, nn_links, amplitudes_phases,
  model_parameter_list)
}
\arguments{
\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{amplitudes_phases}{amplitude and phase of seasonal forcing for each patch.}

\item{model_parameter_list}{list of user-defined model parameters.}
}
\description{
The function creates an equilibrium initialisation state to be used within
later model runs.
}
