% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_init_create.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{equilibrium_init_create}
\usage{
equilibrium_init_create(agec, death, nearest_neighbor_patches,
  model_parameter_list)
}
\description{
\code{equilibrium_init_create} creates an equilibrium initialisation
state to be used within later model runs
}
