% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_init_create.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{Generate equilibrium state}
\usage{
equilibrium_init_create(agec, death, nn_links, amplitudes_phases,
  vaccine_age = NULL, model_parameter_list)
}
\arguments{
\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{amplitudes_phases}{Amplitude and phase of seasonal forcing for each patch.}

\item{vaccine_age}{Vector of binary indicators for age groups vaccination.
1 = vaccinate, 0 = do not vaccinate. Same length as \code{agec}. Default = NULL.}

\item{model_parameter_list}{List of user-defined model parameters.}
}
\description{
The function creates an equilibrium initialisation state to be used within
later model runs.
}
