% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_init_create.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{equilibrium_init_create}
\usage{
equilibrium_init_create(agec, death, nn_links, model_parameter_list)
}
\arguments{
\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{model_parameter_list}{list of user-defined model parameters.}

\item{age}{Vector of age group widths.}
}
\description{
\code{equilibrium_init_create} creates an equilibrium initialisation
state to be used within later model runs.
}
