% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_H}
\alias{format_output_H}
\title{Format human model outputs}
\usage{
format_output_H(x, var_select = NULL, keep = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of human compartment names, \code{c("S", "I1", "R1")}. In
  addition a number of additional variables can be requested. These include:
\itemize{
      \item{"Ntotal"}{ Total number of people }
      \item{"inf_1"}{ Daily infections }
      \item{"MC"}{ Daily microcephaly cases }
      \item{"inf_1_w"}{ Weekly infections per 1000 }
      \item{"MC_w"}{ Weekly microcephaly cases per 1000 }
      }}

\item{keep}{name of variable to stratify by
(allowed are \code{c("patch", "vaccine")}. Default is no stratification)}
}
\value{
Formatted long data.frame
}
\description{
The function formats human-realetd model outputs as a long data frame.
}
