% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_M}
\alias{format_output_M}
\title{Format mosquito model outputs}
\usage{
format_output_M(x, var_select = NULL, keep = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of mosquito compartment names,
e.g. \code{c("Lwt", "Mwt_S", "Mwb_E1")}, or mosquito-related variables,
e.g. \code{c("Kc", "Delta", "R0t_1")}.}

\item{keep}{name of variable to stratify by
(only allowed \code{"patch"}. Default is no stratification)}
}
\value{
Formatted long data.frame
}
\description{
The function formats mosquito-related model outputs as a long data frame.
}
