% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{lag_diff}
\alias{lag_diff}
\title{Calculate lagged differences}
\usage{
lag_diff(x, lag)
}
\arguments{
\item{x}{vector or matrix of numeric values. In the case of a matrix,
the rows indicate the time-varying observations.}

\item{lag}{the lag value (numeric)}
}
\description{
The function returns a vector of lagged difference values using \code{diff}
with initial padding
}
