% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lag_diff}
\alias{lag_diff}
\title{\code{lag_diff} returns a vector of lagged difference values with initial padding.}
\usage{
lag_diff(x, lag)
}
\arguments{
\item{x}{vector of numeric values}

\item{lag}{the lag value (numeric)}
}
\description{
\code{lag_diff} returns a vector of lagged difference values with initial padding.
}
