% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_parameters_list.R
\name{model_param_list_create}
\alias{model_param_list_create}
\title{\code{model_param_list_create} creates a list of user-defined
parameters.}
\usage{
model_param_list_create(YL = 364, DT = 1, NP = 21, age_per = 1,
  N_human = 3e+07, Omega = 1, DeltaBase = 0.2, Sigma = 0.025,
  Epsilon = 1/19, Rm = 2.69, Mwt_mean = 1.5, Mwt_cv = 0.15,
  eip_mean = 8.4, incub = 5.5, inf_per = 6, nu = 1/21,
  Kappa = 0.5, Beta_hm_1 = 0.7, Beta_mh_1 = 0.7, season,
  propMwtControl = 0, TimeMwtControlOn = 1.5,
  TimeMwtControlOff = 2.5, Wb_cyto = 1, Wb_mat = 1, Wb_fM = 0.95,
  Wb_fF = 0.95, Wb_relsusc1 = 0.9, Wb_relinf1 = 0.5,
  Wb_starttime = 1, Wb_introlevel = 0, Wb_introduration = 60,
  vacc_cu_minage = 2, vacc_cu_maxage = 15, vacc_cu_coverage = 0.7,
  vacc_cu_time = 150, vacc_child_age = 1, vacc_child_coverage = 0.56,
  vacc_child_starttime = 150, vacc_child_stoptime = 150,
  dis_pri = 0.2, rho_prim = 1, phi_prim = 1, vacceff_prim = 1,
  other_foi = 0.025, other_prop_immune = 0, propTransGlobal = 5e-04,
  propTransNN = 0.005, BG_FOI = 1e-08, AGE_REC = 2,
  PropDiseaseReported = 1)
}
\arguments{
\item{YL}{Duration of a calendar year. Default = 364.}

\item{DT}{Time step size. Default = 1.}

\item{NP}{Number of patches. Default = 21.}

\item{age_per}{Time (weeks) between age updates. Default = 1.}

\item{N_human}{Human population in each patch. Default = 30000000.}

\item{Omega}{Intensity of density dependence of mosquito larvae mortality rate.
0 = no density dependence. Default = 1.}

\item{DeltaBase}{Adult mosquito mortality rate. Default = 0.2.}

\item{Sigma}{Larval mosquito mortality rate. Default = 0.025}

\item{Epsilon}{Larval mean development rate (1 / larvae mean development time in days). Default = 1/19.}

\item{Rm}{Mosquito reproduction number (based on adult female fecundity and mortality rates). Default = 2.69.}

\item{Mwt_mean}{Mean of the lognorm distribution of the at-equilibrium number of adult female mosquitos per person, per patch, without seasonality. Default = 1.5.}

\item{Mwt_cv}{Standard deviation of the lognorm distribution of the at-equilibrium number of adult female mosquitos per person, per patch, without seasonality. Default = 0.15}

\item{eip_mean}{Mean Extrinsic Incubation Period (days). Default = 8.4.}

\item{incub}{Intrinsic Incubation Period (days). Default = 5.5.}

\item{inf_per}{Total duration of human infectiousness (days). Default = 6.}

\item{nu}{Inverse of virus generation time (1 / days). Virus generation time = serial interval. Default = 1 / 21.}

\item{Kappa}{Mosquito biting rate. Default = 0.5.}

\item{Beta_hm_1}{Per bite transmission probability from humans to mosquitoes. Default = 0.7.}

\item{Beta_mh_1}{Per bite transmission probability from mosquitoes to humans. Default = 0.7.}

\item{season}{Logical for controlling the effect of seasonality.}

\item{propMwtControl}{Proportion of adult wild type mosquitoes which are killed by control. Default = 0.}

\item{TimeMwtControlOn}{Year of start of control of adult wild type mosquitoes. Default = 4.}

\item{TimeMwtControlOff}{Year of end of control of adult wild type mosquitoes. Default = 5.}

\item{Wb_cyto}{Degree of cytoplasmic incompatibility induced by Wolbachia. Default = 1.}

\item{Wb_mat}{Degree of vertical transmission of Wolbachia. Default = 1.}

\item{Wb_fM}{Increase in mortality induced by Wolbachia. Default = 0.95.}

\item{Wb_fF}{Redution in fecundity induced by Wolbachia. Default = 0.95.}

\item{Wb_relsusc1}{Infectivity of a human host to Wolbachia infected mosquitoes
(time τ after host infection). Default = 0.9.}

\item{Wb_relinf1}{Infectiousness of Wolbachia-infected mosquitoes time τ after infection. Default = 0.75.}

\item{Wb_starttime}{Time of first release of Wolbachia-infected mosquiotes (years). Default = 1.}

\item{Wb_introlevel}{Ratio of Wolbachia-infected to wild type mosquitoes at introduction.
(Not the proportion of Wolbachia AFTER introduction). Default = 0.}

\item{Wb_introduration}{Duration of Wolbachia release (days). Default = 60.}

\item{vacc_cu_minage}{Default = 2.}

\item{vacc_cu_maxage}{Default = 15.}

\item{vacc_cu_coverage}{Default = 0.7.}

\item{vacc_cu_time}{Default = 100.}

\item{vacc_child_age}{Deafult = 3.}

\item{vacc_child_coverage}{Default = 0.75.}

\item{vacc_child_starttime}{Default = 30.}

\item{vacc_child_stoptime}{Default = 30.}

\item{dis_pri}{Proportion of infections which are symptomatic. Default = 0.2.}

\item{rho_prim}{Default = 1.}

\item{phi_prim}{Default = 1.}

\item{vacceff_prim}{Efficacy of vaccination in reducing infection. Default = 0.3.}

\item{other_foi}{Default = 0.025.}

\item{other_prop_immune}{Default = 0.}

\item{propTransGlobal}{Proportion of transmission between all patches. Default = 0.0005.}

\item{propTransNN}{Proportion of transmssion with nearest-neighbor patches. Default = 0.005.}

\item{BG_FOI}{Background FOI. Default = 10e-8.}

\item{AGE_REC}{Default = 2.}

\item{PropDiseaseReported}{Reporting rate of symptomatic cases. Default = 0.1.}
}
\description{
\code{model_param_list_create} creates a list of user-defined
parameters.
}
