% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{plot.Zika_model_simulation}
\alias{plot.Zika_model_simulation}
\title{Plot human states}
\usage{
\method{plot}{Zika_model_simulation}(x, var_select = NULL, summarise = NULL)
}
\arguments{
\item{x}{Zika_model_simulation object.}

\item{var_select}{Vector of variable names to plot (default is all)}

\item{summarise}{name of variables by which to stratify (default is no stratification)}
}
\description{
The function makes a plot of the model human compartments.
}
