% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compartments.R
\name{plot_compartments}
\alias{plot_compartments}
\title{\code{plot_compartments} makes a plot of the model compartments.}
\usage{
plot_compartments(df, compart_names, ttl)
}
\arguments{
\item{df}{The dataframe with the data to plot.}

\item{compart_names}{The full name of the model compartments.}

\item{ttl}{The plot title.}
}
\description{
\code{plot_compartments} makes a plot of the model compartments.
}
