% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compartments.R
\name{plot_compartments}
\alias{plot_compartments}
\title{Plot human states}
\usage{
plot_compartments(df)
}
\arguments{
\item{df}{The dataframe with the data to plot.}
}
\description{
The function makes a plot of the model human compartments.
}
