% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compartments.R
\name{plot_compartments}
\alias{plot_compartments}
\title{\code{plot_compartments} makes a plot of the model human compartments.}
\usage{
plot_compartments(df)
}
\arguments{
\item{df}{The dataframe with the data to plot.}
}
\description{
\code{plot_compartments} makes a plot of the model human compartments.
}
