% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demographics.R
\name{plot_demographics}
\alias{plot_demographics}
\title{Plot demographics}
\usage{
plot_demographics(df)
}
\arguments{
\item{df}{The dataframe with the data to plot.}
}
\description{
The function makes a faceted plot of the model demographics outputs.
}
