% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_demographics}
\alias{plot_demographics}
\title{plot_demographics}
\usage{
plot_demographics(df)
}
\arguments{
\item{df}{The dataframe with the data to plot.}
}
\description{
\code{plot_demographics} makes a faceted plot of the model diagnostics.
}
