% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{plot_Kc_eip_delta}
\alias{plot_Kc_eip_delta}
\title{Plot seasonally-varying variables}
\usage{
plot_Kc_eip_delta(out)
}
\arguments{
\item{out}{The dataframe with the data to plot.}
}
\description{
The function makes plots of the mosquito larvae carrying capacity, virus extrinsic
 incubation period, adult mosquito mortality rate and reproduction number,
 averaged across patches, and arrange them in a grid.
}
