% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_processing}
\alias{post_processing}
\title{Post-process model outputs}
\usage{
post_processing(dat, DT)
}
\arguments{
\item{dat}{list of model outputs from the model run.}
}
\description{
The function processes model outputs to calculate metrics of interest.
}
