% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_processing}
\alias{post_processing}
\title{post_processing}
\usage{
post_processing(dat, out_pth)
}
\arguments{
\item{dat}{list of model outputs from the model run.}

\item{out_pth}{The path where to save the plot}
}
\description{
\code{post_processing} post processes model outputs and saves plots of diagnostics.
}
