% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs and odin model}
\usage{
run_model(agec, death, nn_links, amplitudes_phases, time, season = FALSE)
}
\arguments{
\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{amplitudes_phases}{Amplitude and phase of seasonal forcing for each patch.}

\item{time}{time in days.}

\item{season}{Logical for controlling the effect of seasonality.
TRUE = maximum effect of seasonality.
FALSE = no effect of seasonality. Default = FALSE}
}
\value{
list of model outputs
}
\description{
The function creates an odin generator function and runs an instance of the model
using user-defined parameters and an equilibrium initialisation state.
}
