% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{\code{run_model} runs model using declared agec, death and nn_links}
\usage{
run_model(agec, death, nn_links, time, season = FALSE)
}
\arguments{
\item{agec}{Vector of age group widths.}

\item{death}{Numeric of mortality rates.}

\item{nn_links}{8 nearest-neigbors to each patch.}

\item{time}{time in days.}

\item{season}{Logical for controlling the effect of seasonality.}
}
\description{
\code{run_model} runs model using declared agec, death and nn_links
}
